Split line item updaters
