+++ phoenix-scala/phoenix/app/phoenix/routes/admin/OrderRoutes.scala
@@ -15,7 +15,7 @@ import phoenix.payloads.UpdateShippingMethod
 import phoenix.services.Authenticator.AuthData
 import phoenix.services.carts._
 import phoenix.services.orders._
-import phoenix.services.{Checkout, LineItemUpdater}
+import phoenix.services.Checkout
 import phoenix.utils.aliases._
 import phoenix.utils.apis.Apis
 import phoenix.utils.http.CustomDirectives._
@@ -50,7 +50,7 @@ object OrderRoutes {
           pathPrefix("line-items") {
             (patch & pathEnd & entity(as[Seq[UpdateOrderLineItemsPayload]])) { reqItems ⇒
               mutateOrFailures {
-                LineItemUpdater.updateOrderLineItems(auth.model, reqItems, refNum)
+                OrderLineItemUpdater.updateOrderLineItems(auth.model, reqItems, refNum)
               }
             }
           } ~
@@ -58,7 +58,7 @@ object OrderRoutes {
           pathPrefix("order-line-items") {
             (patch & pathEnd & entity(as[Seq[UpdateOrderLineItemsPayload]])) { reqItems ⇒
               mutateOrFailures {
-                LineItemUpdater.updateOrderLineItems(auth.model, reqItems, refNum)
+                OrderLineItemUpdater.updateOrderLineItems(auth.model, reqItems, refNum)
               }
             }
           } ~
@@ -93,13 +93,13 @@ object OrderRoutes {
           // deprecated in favor of /carts route
           (post & path("line-items") & pathEnd & entity(as[Seq[UpdateLineItemsPayload]])) { reqItems ⇒
             mutateOrFailures {
-              LineItemUpdater.updateQuantitiesOnCart(auth.model, refNum, reqItems)
+              CartLineItemUpdater.updateQuantitiesOnCart(auth.model, refNum, reqItems)
             }
           } ~
           // deprecated in favor of /carts route
           (patch & path("line-items") & pathEnd & entity(as[Seq[UpdateLineItemsPayload]])) { reqItems ⇒
             mutateOrFailures {
-              LineItemUpdater.addQuantitiesOnCart(auth.model, refNum, reqItems)
+              CartLineItemUpdater.addQuantitiesOnCart(auth.model, refNum, reqItems)
             }
           } ~
           // deprecated in favor of /carts route
